Refactor AccountStatus and fix missing device message
What changed:
Since AccountStatus created after we are logged in now there's no point in checking if
we have cookie or not. With this in mind we can now reduce the amount of account states
to just deactivated, newDevice and activated.
