+++ Source/UserSession/ZMUserSession+Private.h
@@ -25,6 +25,7 @@
 @class ManagedObjectContextChangeObserver;
 @class LocalNotificationDispatcher;
 @class ZMStoredLocalNotification;
+@class AccountStatus;
 
 #import "ZMUserSession.h"
 
@@ -47,6 +48,7 @@ NS_ASSUME_NONNULL_BEGIN
 @property (nonatomic, readonly) ZMTransportSession *transportSession;
 @property (nonatomic, readonly) NSManagedObjectContext *searchManagedObjectContext;
 @property (nonatomic, readonly) OperationStatus *operationStatus;
+@property (nonatomic, readonly) AccountStatus *accountStatus;
 @property (nonatomic, readonly) LocalNotificationDispatcher *localNotificationDispatcher;
 @property (nonatomic, nullable) ManagedObjectContextChangeObserver *messageReplyObserver;
 @property (nonatomic, nullable) ManagedObjectContextChangeObserver *likeMesssageObserver;
